+++ src/build.py
@@ -201,6 +201,7 @@ source = Split(
     json/geometry_grammar.cpp
     json/geometry_parser.cpp
     json/feature_grammar.cpp
+    json/feature_parser.cpp
     json/feature_collection_parser.cpp
     json/geojson_generator.cpp
     processed_text.cpp