+++ src/svg/output/process_symbolizers.cpp
@@ -71,7 +71,7 @@ bool svg_renderer<OutputIterator>::process(rule::symbolizers const& syms,
     if (process_path)
     {
         // generate path output for each geometry of the current feature.
-        for(unsigned i=0; i<feature.num_geometries(); ++i)
+        for(std::size_t i=0; i<feature.num_geometries(); ++i)
         {
             geometry_type & geom = feature.get_geometry(i);
             if(geom.size() > 0)