+++ CHANGELOG.md
@@ -8,6 +8,8 @@ For a complete change history, see the git log.
 
 ## Future
 
+- Added the ability to create a mapnik Feature from a geojson feature with `mapnik.Feature.from_geojson` in python.
+
 - Added to python bindings: `has_tiff`, `has_png`, `has_webp`, `has_proj4`, `has_svg_renderer`, and `has_grid_renderer`
 
 - Made it possible to disable compilation of `grid_renderer` with `./configure GRID_RENDERER=False` (#1962)